Adjust styles to handle custom checkbox icons
