--- app-frontend/src/app/components/common/toggle/toggle.component.js
@@ -1,15 +1,0 @@
-import toggleTpl from './toggle.html';
-const rfToggle = {
-    templateUrl: toggleTpl,
-    controller: 'ToggleController',
-    transclude: true,
-    bindings: {
-        value: '<',
-        onChange: '&',
-        radio: '<',
-        className: '@'
-    },
-    controllerAs: '$ctrl'
-};
-
-export default rfToggle;
