--- app-frontend/src/app/components/common/toggle/toggle.controller.js
@@ -1,19 +1,0 @@
-export default class ToggleController {
-    constructor() {
-        'ngInject';
-    }
-    $onInit() {
-        this.currentValue = this.value || false;
-        this.radio = this.radio || false;
-    }
-    $onChanges(changes) {
-        if ('value' in changes) {
-            this.currentValue = changes.value.currentValue;
-        }
-    }
-    toggle($event) {
-        $event.stopPropagation();
-        $event.preventDefault();
-        this.onChange({value: !this.currentValue});
-    }
-}
