+++ app-frontend/src/app/components/common/toggle/toggle.html
@@ -1,7 +1,8 @@
 <div class="checkbox {{$ctrl.className}}"
      ng-class="{
-       'active': $ctrl.currentValue, 
-       'radio': $ctrl.radio
+       'active': $ctrl.currentValue,
+       'radio': $ctrl.radio,
+       'custom-icon': $ctrl.customIcon
      }"
      ng-click="$ctrl.toggle($event)"
      ng-transclude>
