+++ app-frontend/src/app/components/common/toggle/toggle.module.js
@@ -1,10 +1,43 @@
 import angular from 'angular';
-import ToggleComponent from './toggle.component.js';
-import ToggleController from './toggle.controller.js';
+import toggleTpl from './toggle.html';
 
 const ToggleModule = angular.module('components.toggle', []);
 
-ToggleModule.component('rfToggle', ToggleComponent);
+const rfToggle = {
+    templateUrl: toggleTpl,
+    controller: 'ToggleController',
+    transclude: true,
+    bindings: {
+        value: '<',
+        onChange: '&',
+        radio: '<',
+        className: '@',
+        customIcon: '<'
+    },
+    controllerAs: '$ctrl'
+};
+
+class ToggleController {
+    constructor() {
+        'ngInject';
+    }
+    $onInit() {
+        this.currentValue = this.value || false;
+        this.radio = this.radio || false;
+    }
+    $onChanges(changes) {
+        if ('value' in changes) {
+            this.currentValue = changes.value.currentValue;
+        }
+    }
+    toggle($event) {
+        $event.stopPropagation();
+        $event.preventDefault();
+        this.onChange({value: !this.currentValue});
+    }
+}
+
+ToggleModule.component('rfToggle', rfToggle);
 ToggleModule.controller('ToggleController', ToggleController);
 
 export default ToggleModule;
