+++ app-frontend/src/app/components/lab/templateCreateModal/templateCreateModal.html
@@ -53,7 +53,6 @@
         <div class="form-group">
           <label for="name">Make this template publicly available?</label>
           <rf-toggle value="$ctrl.templateIsPublic()" on-change="$ctrl.onVisibilityChange(value)">
-            <i class="icon-check"></i>
           </rf-toggle>
         </div>
         <div ng-if="$ctrl.currentError" class="color-danger text-center form-group">
