+++ app-frontend/src/app/pages/imports/datasources/detail/detail.html
@@ -36,7 +36,6 @@
             <div class="cta-flex-text">
               <div>
                 <rf-toggle value="$ctrl.isPublic" on-change="$ctrl.changeVisibility()">
-                  <i class="icon-check"></i>
                 </rf-toggle>
                 <span title="Public">Public</span>
               </div>
