+++ app-frontend/src/app/pages/projects/edit/annotate/import/import.html
@@ -29,7 +29,6 @@
           </div>
           <div class="list-group-right">
             <rf-toggle value="$ctrl.isMachineData" on-change="$ctrl.isMachineMade()">
-              <i class="icon-check"></i>
             </rf-toggle>
           </div>
         </div>
