+++ app-frontend/src/app/pages/projects/edit/exports/new/new.html
@@ -111,7 +111,6 @@
           <div class="output-selection">
             <rf-toggle value="$ctrl.isCurrentProcessingOption(option)"
                        on-change="$ctrl.handleOptionChange(value, option)">
-              <i class="icon-check"></i>
             </rf-toggle>
           </div>
         </div>
